Move auth to pages
