+++ src/app/pages/auth/auth-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { NoAuthGuard } from '../guards/no-auth.guard';
-import { CentralCardLayoutComponent } from '../layouts/central-card-layout/central-card-layout.component';
+import { NoAuthGuard } from '../../guards/no-auth.guard';
+import { CentralCardLayoutComponent } from '../../layouts/central-card-layout/central-card-layout.component';
 import { LoginComponent } from './login/login.component';
 
 const routes: Routes = [
