+++ src/app/pages/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { LoginComponent } from './login/login.component';
 import { AuthRoutingModule } from './auth-routing.module';
-import { LayoutsModule } from '../layouts/layouts.module';
+import { LayoutsModule } from '../../layouts/layouts.module';
 
 @NgModule({
   declarations: [
